fix(avr): atomic ISR-shared counter access and 32-bit OTA block address
- canHandlerAtmega328P: wrap the intCount decrement in ATOMIC_BLOCK; the
  3-instruction read-modify-write raced with rxInterrupt's increment, and
  a lost increment with the edge-triggered INT line could leave a received
  CAN frame stuck in the MCP2515 forever.
- pumpControl: route every main-loop access to the 16-bit flowCounter
  through atomic read/clear helpers; on AVR the two 8-bit halves could
  tear against the flow ISR, producing phantom FLOW_STUCK/FLOW_OVERRUN
  errors. The flow check now also reads the counter once per evaluation.
  Non-AVR builds (native tests) compile to plain accesses.
- ota: widen the flash block address multiply to 32 bits; AVR's 16-bit
  int wrapped the product for flashBlockNumber >= 2.

Verified: all 7 envs build clean (-Wall -Wextra -Werror), 359/359 native
tests pass, cppcheck + clang-tidy report no defects, and the disassembly
shows SREG-save/cli/restore around every protected access.

+++ lib/ota/src/ota.cpp
@@ -16,7 +16,8 @@ bool OTA::start(uint16_t flashBlockNumber, uint32_t fwSize, uint16_t fwCrc) {
   flash.chipErase();                                              // Attempt to erase the FLASH block.
   this->fwSize = fwSize;                                          // Save FW size.
   this->fwCrc = fwCrc;                                            // Store FW CRC.
-  flashBlockBeginAddress = flashBlockNumber * flashBlockTobytes;
+  // 32-bit multiply: with AVR's 16-bit int the product would wrap for flashBlockNumber >= 2.
+  flashBlockBeginAddress = static_cast<uint32_t>(flashBlockNumber) * flashBlockTobytes;
   flashPointer = 0;                                               // Reset flash pointer.
   otaState = OtaState::START;
   return true;                                                    // Return success.
